refactor: 패키지명의 대문자 제거
